INTEL_DII/FIXME: drm/i915/vm_bind: Unlock the vm_bind lock across relocation processing.
Processing relocations within the vm_bind lock will lead to an
undesirable lock ordering vm_bind_lock -> fault.

Moving forward, a vm with persistent vmas will not support relocations,
and currently for hardware GEN12+ where we intend to support vm_bind,
relocations are deprecated.

Unlock the vm_bind lock across relocation processing. Ideally we should
not be taking the lock at all if the vm is not marked for persistent
vmas, but this will do for now.

FIXME: to be squashed with suitable previous commits before upstreaming.

Reviewed-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Signed-off-by: Thomas Hellström <thomas.hellstrom@linux.intel.com>