INTEL_DII: drm/i915: Allow unaligned buddy reservations
As the goal is simply to reserve a range within the buddy to prevent
future allocations, we can accept a byte range and remove the
overlapping chunk-aligned blocks. This allows the API to remain flexible
(where the caller can simply pass in reserverations determined by HW
irrespective of the buddy initialisation parameters) while maintaining
the chunk aligned properties of the buddy itself. Equally the caller's
range can extend to either size of the buddy range without loss of
generality.

v2: Align reservations to chunk_size to avoid splitting the smallest
blocks; and test it!

v3(Tejas): Fix "end" calculation for range

Signed-off-by: Chris Wilson <chris.p.wilson@intel.com>
Cc: Matthew Auld <matthew.auld@intel.com>
Signed-off-by: Tejas Upadhyay <tejas.upadhyay@intel.com>