INTEL_DII: drm/i915/gt: Keep scheduling information in the context structure
GuC (and execlist as well) track context scheduling parameters
in the engine data structure.

But, since the debugger came along, we started needing some
scheduling parameters more tied to the context rather than the
engine.

For example in the debugger's case we want the preemption timeout
to be '0' for all the debugged contexts. If we rely solely on the
"props" we apply the same values to all contexts executed in that
engine.

On the other hand setting the same parameters runtime, we might
miss what happens in between, for example when the context is
unpinned and then pinned again, the GuC interface restores the
parameters from the "props".

Set up a structure in intel_context that will be the main
placeholder for the context specific scheduling information.

As of now store the preempt_timeout_ms with the
timeslice_duration_ms.

v2: Check for a valid GuC context before updating (JohnH)

Suggested-by: John Harrison <John.C.Harrison@Intel.com>
Signed-off-by: Andi Shyti <andi.shyti@linux.intel.com>
Reviewed-by: John Harrison <John.C.Harrison@Intel.com>