INTEL_DII: drm/i915/debugger: Disable preemption timeout when debugger is active
When a client connects to the debugger disable the preemption
timeout for the context we are debugging.

Execlist and GuC have a different approach when it comes to
handling the preemption timeout:

 * In execlist skip resetting when hitting a preemption timeout
   interrupt.

 * In GuC set the preemption timeout to the context belonging to
   the debugger session.

In the second case we loop through all the active contexts and if
they belong to our debugging session we set their timeout to '0'.
From then all the context created will ask to disable their
preemption. When the debugger is closed we do the opposite and
set the preemption timeout as defined in the
engine->props.preemption_timeout_ms.

The execlist solution has been suggested by Chris Wilson.

v2: Downstream return void from context_apply_all (Pallavi)

Signed-off-by: Andi Shyti <andi.shyti@linux.intel.com>
Cc: Chris Wilson <chris.p.wilson@intel.com>
Cc: Maciej Patelczyk <maciej.patelczyk@intel.com>
Cc: Mika Kuoppala <mika.kuoppala@linux.intel.com>
Reviewed-by: Maciej Patelczyk <maciej.patelczyk@intel.com>