drm/i915: Increase slow timeout for PCode request
During boot we ocassionally notice ETIMEDOUT errors during
initial CDCLK change, which however succeed if we retry a bit
more. So lets be not so pessimistic and increase slow time
out, which means we wait for some time for PCode to accept the
request and be able to communicate again.

v2: Increase timeout also for subsequent set freq request, because
    it can fail same way also.
Signed-off-by: Stanislav Lisovskiy <stanislav.lisovskiy@intel.com>
Acked-by: Matt Roper <matthew.d.roper@intel.com>