drm/i915/guc: Fix GuC relay logging debugfs failing on open
When GuC-Error-Capture was introduced, we created
buf_in_use as a way to identify if relay logging
had started. It is meant to replace the previous
method where a mmap of the GuC log buffer was
the indicator but not since GuC Error Capture
shares that mapping throughout operation.

However, that method of checking was not updated
when the debugfs guc_log_relay_ctl_open was called.
Fix that check.

Fixes:
   drm/i915/guc: Add capture region into intel_guc_log

Signed-off-by: Alan Previn <alan.previn.teres.alexis@intel.com>
Reviewed-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Link: https://patchwork.freedesktop.org/patch/479021/?series=101603&rev=1