INTEL_DII: drm/i915/gt: Add driver-FLR capability at gt start/end
Depending on the platform configuration, do a Driver-FLR after
MMIO-init during gt's init (and again later at gt's fini).

The driver-FLR is not the same as PCI config space FLR in that
it doesn't reset the SGUnit and doesn't modify the PCI config
space. Thus, it doesn't require a re-enumeration of the PCI BARs.
However, the driver-FLR does cause a memory wipe of graphics memory
on all discrete GPU platforms or a wipe limited to stolen memory
on the integrated GPU platforms.

Driver-FLR was designed into the architecture as a way to reset the
KCR / GSC microcontroller (and firwmware) if required.

v2: Change driver_flr param to intel_gt (Piotr)

Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Signed-off-by: Alan Previn <alan.previn.teres.alexis@intel.com>
Reviewed-by: Radhakrishna Sripada <radhakrishna.sripada@intel.com>