INTEL_DII/NOT_UPSTREAM: Add mechanism for prelim query calls
Add a mechanism to support future query APIs before the have been
upstreamed. The intention is the mechanism stays in DII. Individual
queries are initially added as PRELIM but eventually get pushed
upstream. Once they are official and have been backported and all UMDs
have been updated to match, the PRELIM version would be dropped.
Meanwhile, this patch stays ready for the next batch of new queries to
be added.

Signed-off-by: John Harrison <John.C.Harrison@Intel.com>
Reviewed-by: Ashutosh Dixit <ashutosh.dixit@intel.com>