INTEL_DII: drm/i915/pf: Expose the tile-mask to rest of the driver
VFs can be provisioned in a way where subset of tiles ends up being
available for VF use. In this case, only a subset of tiles is
interesting for migration.
Expose the tile-mask related helpers to the rest of the driver so that
it can be exported to VFIO driver later on.

v2: add proper locking in-place (MichalWa)

Signed-off-by: Michał Winiarski <michal.winiarski@intel.com>
Acked-by: Michal Wajdeczko <michal.wajdeczko@intel.com>