drm/i915/gt: Push srcu synchronisation point outside of reset timeout
srcu synchronisation, even in expedited form, can still be starved by a
system under stress to the point where including it in our watchdog
(which forces the device to be declared wedged upon timeout) is
problematic.

The requirement for our srcu synchronisation is that it is before the
actual GT reset (to ensure no user is still accessing the HW during the
reset) and after we have primed the backoff to prevent new users from
starting their access, with the primary interaction being with
vm_fault_gtt(). It's current position was to be as late as possible, but
that pushes it under the reset timeout, but may incur delays unrelated
to the driver and resetting the device -- causing inappropriate wedging.

Move the scru synchronisation point to outside the timeout, and place it
in the same function as the rest of the scru backoff so the sequence and
interaction of the rcu/srcu is clearly (and more consistent with the
srcu guide).

Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Reviewed-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>