INTEL_DII: drm/i915/sriov: Add reset locking to post migration
While the post-migration callback is working, any GT reset should be
discarded - we need a complete post migration recovery and not reset. For
engine reset, we do not have to care as it is both triggered and handled on
GuC side.

There already is a mechanism for preventing reset within the driver - it's
I915_RESET_BACKOFF flag, with accompanying backoff_srcu. It matches to our
scenario not only in how it is implemented, but in the idea behind it as
well. Migration can be thought of as a special kind of reset, more extensive
in scope than normal GT reset.

Setting I915_RESET_BACKOFF flag causes any further GT resets to be discarded.

Signed-off-by: Tomasz Lis <tomasz.lis@intel.com>
Acked-by: Michal Wajdeczko <michal.wajdeczko@intel.com>