INTEL_DII: drm/i915/sriov: Post-migration shutdown/kickstart
After migration, we end up running on another VF, which went through FLR
(among other operations) before the VM was resumed. We need to shut down
all the operations and data which relate to hardware state of the previous VF.

The patch implements shutdown/kickstart functions, which reuse the routines
created for suspend/resume to deal with the environment change caused by
migration.

As we want to continue execution of pending requests after the
migration, we should not flush workloads or disable PPGTT mappings.
This is the reason for detouring from the suspend path and calling
`i915_gem_drain_freed_objects()` instead of `i915_gem_suspend()`.

Co-authored-by: Piotr Piórkowski <piotr.piorkowski@intel.com>
Signed-off-by: Tomasz Lis <tomasz.lis@intel.com>
Acked-by: Michal Wajdeczko <michal.wajdeczko@intel.com>