INTEL_DII: drm/i915/sriov: Handshake GuC for VF config query
While recovering from migration, VF driver need to reestablish communication
with GuC. Depending on where the VM got migrated, this may be different GuC
than the one used before. VF driver needs to query GuC for configuration of
newly assigned VF device, as it is needed to complete the post-migration
recovery.

Only MMIO communication is used for the recovery. Proper CTB communication
is established later, when we kickstart the VF driver back to work.

Signed-off-by: Tomasz Lis <tomasz.lis@intel.com>
Acked-by: Michal Wajdeczko <michal.wajdeczko@intel.com>