INTEL_DII: drm/i915/mtl: Enable Idle Messaging for GSC CS
We are seeing an issue in MTL where Media is unable to go
to C6 because GCS CS idle messaging is disabled. BIOS is enabling
the idle messaging, but they are not forcewaking the GSC before
doing so. We are not certain if this is a HW or BIOS issue for now.
Setting this in the driver to unblock Media C6 testing. This patch
should be reverted when BIOS is fixed.

JIRA: VLK-36665

v2: Move this code to engine_init (Daniele)
v3: Remove duplicate code
v4: Fixed checkpatch warnings (Badal)

This patch was reverted as it caused some problems. However the problems
were isolated to VPU/IPU being disabled in the bios.

Cc: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Reviewed-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Signed-off-by: Vinay Belgaumkar <vinay.belgaumkar@intel.com>