INTEL_DII/NOT_UPSTREAM: mm: Expose NUMA rebalancing via sysctl for debugging
Originally mbind(MPOL_MF_LAZY) was conceived for triggering autonuma
rebalancing on the process, but this was dropped as the only user at the
time was for testing. Testing such as we want to perform to reproduces
problems with numa rebalancing of compute tasks. Copy the method used
for triggering THP compaction (vm/compact_memory) and introduce a
sysctl that we can use in our testing to reproduce the deadlocks.

Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>