INTEL_DII/FIXME: drm/i915/guc: Allow for very slow GuC loading
A failure to load the GuC was observed once on ATS-M where the GuC log
actually showed that the GuC had loaded just fine. The implication
being that the load just took ever so slightly longer than the 200ms
timeout. Given that the actual time should be tens of milliseconds at
the slowest, this is concerning. So increase the timeout and add some
reporting on slow loads.

FIXME: This patch can be completely dropped once the real issue is
determined and fixed.

v2: Bump the timeout even more (there have been instances of DG2
taking over 10s to load) and add a warning if the load is over 200ms
so that the issue can be tracked in CI.
v3: Report the clock frequency before/during load.

Jira: VLK-36744

Signed-off-by: John Harrison <John.C.Harrison@Intel.com>
Reviewed-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>