INTEL_DII/UAPI: PRELIM_UAPI for DG2 compression
DG2 compression support has landed upstream and been backported.  We now
just need a temporary commit to add the old prelim uapi defines.

Note that we previously had some additional prelim uapi aliases:
  PRELIM_I915_FORMAT_MOD_F_TILED_DG2_RC_CCS ==
  PRELIM_I915_FORMAT_MOD_4_TILED_DG2_RC_CCS
and
  PRELIM_I915_FORMAT_MOD_F_TILED_DG2_MC_CCS ==
  PRELIM_I915_FORMAT_MOD_4_TILED_DG2_MC_CCS
and
  PRELIM_I915_FORMAT_MOD_F_TILED_DG2_RC_CCS_CC ==
  PRELIM_I915_FORMAT_MOD_4_TILED_DG2_RC_CCS_CC

The "F" form is dropped in favor of the "4" form since that matches what
we settled on upstream and we don't want the "F" terminology to be
cargo-culted to future platforms.  Dropping these aliases has no impact
on ABI; it will only impact userspace at compile time if/when they
choose to re-sync their own definitions with the latest KMD headers.

Signed-off-by: Matt Roper <matthew.d.roper@intel.com>