drm/i915/gem: Mark userptr vma as special
When we bind user pages to the GPU they become active with side-effects
and must be flushed from the GPU before they can be reused by the
system, causing indefinite system-wide stalls. Anytime the user page
tables for that range are touched, a mmu_notifier will try to invalidate
the shadow PTE (the GPU's binding to those pages) so that they can be
refetched post-change (and so checked against the new user PTE).

The kernel performs a number of background opportunistic tests to try
and defragment memory or to move memory closer to the active node. While
the GPU is using the pages, they cannot be moed, so prevent the
opportunistc migration by marking the vma as special. But we cannot
mark the vma as too special as we need to continue using it as a plain
struct page area.

Testcase: igt/gem_userptr_blits/numa-balance
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Acked-by: Mika Kuoppala <mika.kuoppala@linux.intel.com>