INTEL_DII/FIXME: drm/i915: Allow non-persistence without resets
In order to debug some issues, it is useful to be able to run with
resets disabled. Compute workloads must run with persistence disabled.
However, i915 refuses to allow non-persistent contexts when resets are
disabled due to persistence being unnecessarily tied to heartbeats,
resets and all sorts of other things. This makes it impossible to
debug some issues.

So add a module parameter to override the anti-reset check when
disabling persistence.

FIXME: Once the planned re-write of context peristence is completed,
this will no longer be necessary and the patch can be dropped.

Jira: XDEPS-2687

Signed-off-by: John Harrison <John.C.Harrison@Intel.com>
Reviewed-by: Umesh Nerlige Ramappa <umesh.nerlige.ramappa@intel.com>