drm/i915/gt: Wait for an available fence
Under memory management pressure, we may be trying to fault in more
processes simultaneously than the HW has available detiling registers.
Currently, this results in a SIGBUS being sent to a random application
due to the HW exhaustion, which is unfortunate. Add a waitqueue for the
detilers so that we can wait in the pagefault handler until a detiling
register becomes available.

Suggested-by: Matthew Auld <matthew.auld@intel.com>
Closes: https://gitlab.freedesktop.org/drm/intel/-/issues/5161
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Cc: Matthew Auld <matthew.auld@intel.com>