INTEL_DII/NOT_UPSTREAM: drm/i915/guc: dump the EIP on load timeout
We sporadically observe the GuC load failing without any error reported
by the FW and with the status registers indicating that the FW is
stuck in the middle of the load. To better understand where exactly the
GuC is stuck, the GuC FW team has requested that we dump the instruction
pointers of the microcontroller so they can compare that with the code.

Note that on gen 12.5 and newer we need to explicitly enable this debug
capability, while in older gens it is always on.

Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Reviewed-by: Stuart Summers <stuart.summers@intel.com>