INTEL_DII/NOT_UPSTREAM: drm/i915/hwmon: Expose power1_rated_max for < 5.10 kernels
The hwmon attribute power1_rated_max was introduced in the 5.10 kernel and
is not available when DII is backported to < 5.10 kernels. To circumvent
this, create a custom power1_rated_max sysfs attribute for < 5.10
kernels. This allows us to exposed a uniform interface to userspace across
all kernels.

It also allows us to discontinue the power1_max_default sysfs file which
was serving this purpose for < 5.10 kernels in the future after UMD's have
transitioned from power1_max_default to power1_rated_max.

V2: Align with prototype update of "hwm_field_read_and_scale" <Naveena>

Cc: Badal Nilawar <badal.nilawar@intel.com>
Cc: Ayaz Siddiqui <ayaz.siddiqui@intel.com>
Signed-off-by: Ashutosh Dixit <ashutosh.dixit@intel.com>
Signed-off-by: Naveena Kolanupaka <kolanupaka.naveena@intel.com>