INTEL_DII: i915/perf: Use LMEM for noa_wait if available
On 2T PVC, iaf driver init puts pressure on the PCIe bus. When noa wait
bo is allocated outside the gt, the batch that executes the noa_delay
runs much slower and the actual delay is more than double the intended
noa_programming_delay. Use LMEM, if available, to resolves the issue.

Test: i915_selftest@live@perf@live_noa_delay

Signed-off-by: Umesh Nerlige Ramappa <umesh.nerlige.ramappa@intel.com>
Reviewed-by: Ashutosh Dixit <ashutosh.dixit@intel.com>