INTEL_DII: drm/i915/pvc: RC6 WA Increase forcewake ack timeout
Wa_16017528748:

On PVC Bx, it is observed that T2T MDFI counters are moving up
without any traffic after boot or rc6 exit.

When GT is either coming out of reset or RC6 exit, T2T MDFI
controller can inject spurious packets to remote tile.
This spurious packet reachs GT on remote tile causing issues
in GT and resulting MCA.

The proposed WA in pcode at RC6 exit for the above issue:
Pcode has to gate “GT clocks” inside T2T MDFI controller
when GT is either coming out of reset or RC6 exit.
Pcode has to master TAP (JTAG) network to gate clocks.
Pcode needs to handshake with CSC and keep it in reset
when it is mastering TAP (JTAG) network.
The above Pcode WA is taking around ~50ms in worst case to
complete the handshake.
When CSC traces are enabled in debug mode, the WA is taking
around ~300ms.

Forcewake is blanketed in above Pcode WA sequence to avoid
intermittent rc6 entries.
So, to counteract above forcewake hold period in Pcode WA
at rc6 exit, increased forcewake ack timeout to 500 msecs.

_wait_for_atomic is used, it is a busy wait, In the unlikely scenario
that this call comes from an interrupt handling and RC6 is enabled
we might wait this long timeout in the worst case and cause an impact
in the whole system, including other PVCs in the system. However this
might not happen unless we are already in an error kind of situation
since the  Wa_16015496043 is already blocking RC6 whenever we have
any client connected.
Signed-off-by: Tilak Tangudu <tilak.tangudu@intel.com>
Acked-by: Rodrigo Vivi <rodrigo.vivi@intel.com>
Reviewed-by: Vinay Belgaumkar <vinay.belgaumkar@intel.com>