INTEL_DII/FIXME: drm/i915/gt: Record indirect context w/a into golden default context state
When we first load the proto contexts (before we have defined the golden
default context state), we inhibit context restores. This also inhibits
the wa_bb that will be run as part of the context restore, thus we
currently do not capture any indirect w/a into the default context
state. For completeness, we do want that default state to include
applicable w/a, so after first loading the proto context and uploading
the known good state for the first context, do a couple of context
switches (away from and back to the first contxt) so that we can record
the outcome of the indirect context w/a.

v2:
   - Added missing intel_context_get(ce) - Chris
   - FIXME newly added API doc - Stuart
v4: Update commit message - Matt/Stuart
v5:
   - Add FIXME for API doc - Stuart
   - Manually pin context - Chris
v6: Add comment about intel_context_get(ce) - Stuart
v7: Add FIXME comment in patch itself - Stuart
v8: Add debug with each engine - Lucas

Acked-by: Stuart Summers <stuart.summers@intel.com>
Signed-off-by: Chris Wilson <chris.p.wilson@linux.intel.com>