INTEL_DII/FIXME/UAPI: drm/i915/gt: Add sysfs reset interface per GT
Expose the reset functionality as an API. The user will have the
possibility to reset the GT's by writing any number greater than
'0' to the reset interface.

The reset can be performed per GT and the interface would be
accessible under

/sys/class/drm/cardX/gt/gtX/prelim_reset

or per GPU, in the sense that all the GTs are to be reset. The
latter is accessible in:

/sys/class/drm/cardX/prelim_reset_all_gt

The interface is write only: when writing '1', 'y', 'Y', 'on',
'ON' or 'On' a reset will be triggered. While when writing '0',
'n', 'N', 'off', 'OFF' or 'Off' nothing will happen. Any other
value will return an EINVAL to the user.

Signed-off-by: Andi Shyti <andi.shyti@linux.intel.com>
Acked-by: Ashutosh Dixit <ashutosh.dixit@intel.com>
Acked-by: Christoph Manszewski <christoph.manszewski@intel.com>
Reviewed-by: Chris Wilson <chris.p.wilson@linux.intel.com>